some slight changes to slug creation

+++ src/utils/convertUnusableCodec.js
@@ -10,23 +10,30 @@ fluentFfmpeg.setFfmpegPath(Ffmpeg.path)
  * @param {string} filePath 
  */
 const convertVideo = (filePath)=>{
-    const acceptedVideoExtensions = ['mp4', 'avi', 'mov', 'wmv', 'mkv', 'flv']
-    let filewithoutExtension = filePath.split('.')[0]
-    fluentFfmpeg(`${filePath}`,{timeout:4200}).output(`${filewithoutExtension}.mp4`).on("end",()=>{
-        console.log("done")
-        fs.unlink(path.join(__dirname,filePath))
-        //add code here to delete the mkv or other file
-    })
-    .on("progress",(progress)=>{
-        console.log(progress)
-    })
-    .on("error",(err,stdout,stderr)=>{
-        console.log(err)
-        console.log(stderr)
-        console.log(stdout)
-    }).run()
+    try {
+        const acceptedVideoExtensions = ['mp4', 'avi', 'mov', 'wmv', 'mkv', 'flv']
+        const lastDotOccurence = filePath.lastIndexOf(".")
+        let fileExtension = filePath.slice(lastDotOccurence + 1)
+        let filewithoutExtension = filePath.slice(0,lastDotOccurence)
+        if (acceptedVideoExtensions.includes(fileExtension)) {
+            fluentFfmpeg(`${filePath}`,{timeout:4200}).output(`${filewithoutExtension}.mp4`).on("end",()=>{
+                console.log("done")
+                fs.unlink(path.join(__dirname,filePath))
+                //add code here to delete the mkv or other file
+            })
+            .on("progress",(progress)=>{
+                console.log(progress)
+            })
+            .on("error",(err,stdout,stderr)=>{
+                console.log(err)
+                console.log(stderr)
+                console.log(stdout)
+            }).run()   
+        }
+    } catch (error) {
+        console.log(error)
+    }
 }
-
 
 module.exports = {
     convertVideo
